tasks/src/generate-assets/generate-geosjon/requirements.txt
functions_framework==3.5.0
google-cloud-storage==2.16.0
pyproj==3.6.1
python-dotenv==1.0.1
Requests==2.31.0
google-cloud-bigquery==3.20.1
geojson==3.1.0